requirements-test.txt
pytest
pytest-asyncio
httpx
fastapi
Pillow
